Add dependency: HDF5
Purpose: HDF5 will be used to read RPA-input files.

Why not using JLD2? So far I could not figure out how to read attributes using JLD2.

Notice: Julia-Version is downgraded in this commit. This needs to be fixed later on! For this moment the focus is on new features rather than software concistency.