Ability to use subset of bosonic freqs from chi_0
One can now select the number of positive bosonic frequencies to be used for a calculation. The number of positive bosonic frequencies can be requested via a new input parameter in the configuration file "n_pos_bose_freqs". Obsiously the chi_0 file has to contain at least n_pos_bose_freqs bosonic frequencies (throw error if this is not the case...).